requirements.txt
rio-ui>=0.1.0
sqlalchemy>=1.4.0
pymysql>=1.0.0
python-dotenv>=0.19.0
pandas>=2.0.0